Replace boost::regex with std::regex (or simpler alternatives)
This removes all usage of boost::regex from the code base, mostly
replacing it with std::regex (and in a couple places where regex use was
stupid, replacing it with a more intelligent implementation).

Many of the remaining regexes are quite terrible and I fixed some of
them to be more sensible.  One such fix was the conflate the ipv6 and
non-ipv6 URL regex parsing into a single regex (the only difference
between them is that the ipv6 version allows [IPliteral].

It also eliminates an epee monstrosity that used a macro to induce an
lock around static local variable initialization.  This is retarded:
C++11 *already* guarantees thread-safe local static variable
initialization, and since this didn't eliminate the static local
variable that means you get epee wrapping itself to do what a C++
compiler *ALREADY* has to do anyway inside the epee monstrosity.