Eliminate cryptonote::blobdata typedef
Using a separate type name to describe the purpose of the *variable*
rather than the purpose of the *type* is just wrong.

This is not a typedef that can be practically changed, since everything
just treats it as a string, but it obfuscates the understanding of the
code with no gain: when you look at a function and see that you have a
`cryptonote::blobdata` you have no idea what you can do with it, until
you look up the typedef and see that it is a std::string.  But that
hasn't gained anything: neither type safety nor a typedef that could
ever change.