Remove useless junk from miner: autodetect, miner_conf.json
Autodetect is not that useful in the first place, even if we weren't
PoS, but is especially useless now.

miner_conf.json is this probably-never-used feature to load some extra
crap into the tx_extra.  Just kill it.

Remove the ability (and RPC endpoint) to print the hashrate periodically
to stdout because that is pretty dumb.