Remove useless GET_TRANSACTION_POOL_BACKLOG endpoint
Not only is this endpoint not used anywhere, but it returns (invalid)
cancerous json output when invoked in current stable oxend:

    {
      "jsonrpc": "2.0",
      "id": "0",
      "result": {
        "backlog": "l\u0007\u0000\u0000\u0000\u0000\u0000\u0000�\u001f�\u0000\u0000\u0000\u0000\u0000Y�������",
        "status": "OK",
        "untrusted": false
      }
    }

where that cancer in the backlog field is the raw bytes copied out from
under a struct with 3 uint64_t's.  (It's also invalid json because those
unprintable characters are \xff, which isn't valid UTF-8 encoding, and
thus isn't valid JSON).

Just nuke the whole thing since nothing uses it (probably because it is
such cancer).