Use null contract address in HF20 uptime proof BLS sigs
Testnet HF20 uptime proofs failed to send because the testnet contract
address was empty, and so the BLS signature would fail because the tag
hash would fail to parse the empty hex contract address into an
eth::address.

This changes the HF20 uptime proof BLS signature to use a null eth
address instead of the contract address for these signatures.

This bug was caused by a recent change that changed the "TODO FIXME"
0x0000... null address in the network_config to an empty string, but in
doing so exposed a more serious problem: because that network_config
value is going to change between the initial HF20 and final HF21
releases, uptime proofs generated *during HF20* by the HF20 release and
the HF21-ready release would not have been compatible because they would
have used a different contract address in the signature.

This forces the use of the null address always during HF20 (regardless
of the actual network_config contract address) to solve both problems.