Remove stagenet db on snapshot package update
Because the default config specifies a data-dir (rather than using the
default), we will end up trying to use the existing one rather than a
new one for the deb version of the package.

This adds upgrade code to the oxend-snapshot postinst to wipe that
database if it exists, when upgrading from a previous version of the
deb.