Fix testnet syncing
Testnet forked on WIP code that didn't yet have the
final-hf20-block-reward-payout, and as a result the current code cannot
sync beyond the second-last block of HF20.  This fixes it with a hack to
bypass that reward payout on testnet (along with a static assert so that
whenever we reboot testnet, we'll fail to compile with the hack still in
place).