Use larger (25 node) pulse quorums starting at HF23
This adds various code to support up to 25-node pulse quorums (with a
16/24 validating quorum signature requirement) starting at HF 23.

The purpose of this extension is to address some moderate weaknesses
that were identified by researchers where a potential malicious large
operator (e.g. controlling at least 25% of the network) could
theoretically wait until they get lucky with a pulse quorum, use it to
influence the the testing quorum, and use the testing quorum to
deregister a handful of nodes.  Repeated many times, this could allow
such a malicious operator to increase their control on the network.

Such an attack would, however, be highly visible (with sudden,
unexpected deregistrations), and would take several months to get lucky
enough to find even one block at which to start the attack, and so while
we feel the threat of this is low, increasing the pulse quorum size is
relatively easy and largely neutralizes the attack.  (The other remedy,
which will follow this commit, solves a related issue which also makes
the attack much more difficult).

This increased quorum size will somewhat increase Oxen blockchain growth
over time: the 7 pulse validator signatures per block *currently* amount
to just under 120MB of blockchain growth per year; switching to 16
signatures per block will increase the blockchain growth rate by another
150MB per year, which we don't feel poses a burden for SN operators.

The idea is that the pulse quorum size is now determined by the number
of active nodes at the block before the quorum: for 25 or more active
SNs, the pulse quorum will be full size (25 nodes); mainnet keeps its
current minimum of 50 nodes for pulse (and so mainnet HF23+ quorums will
*always* be 25 nodes); for smaller number of active nodes (such as often
occurs on testnet) we scale down to whatever we can support (minimum 8).

This also increases the validator signature requirement, slightly, to >=
2/3 of the validator count (i.e. 16 of 24 for full quorums).  Currently
we require 7/11 (~64%).  I believe the thinking was that 7/11 is
actually 8/12 because the block producer implicitly votes by providing
the block, but that isn't exactly right because a malicious quorum of 7
could throw away the producer's block and substitute their own.

For entropy, we use the same pulse entropy process as before, but extend
the selection further back to allow for larger quorums.  Where
previously we were using the pulse random data in H-21 (H = current
height) to select the block producer, and pulse random data in H-20
through H-10 for the 11 validators, we now H-10, back to H-(9+N) to
select the quorum (where the earliest one selects the producer, and the
later one selects validators).  Thus a 12-quorum selection remains
identical.

Details:
- Adds hf23 constant
- Reworks validator bitsets in many, many places to be a `bitset_t` (=
  uint32_t) instead of a uint16_t, so that we can go beyond 16
  validators.
- various pulse "constants" are no longer constants, but now rather
  depend on the HF and number of active service nodes.
- Add block pulse data workaround to encode validator participation in
  HF23 blocks as a varint rather than raw little-endian uint16_t value.
- Add quorumnet connection matrices for quorums of size 21-25, and
  de-clang-format the existing values as clang-format destroyed their
  nice square matrix form.
- Reworked how the pulse_entropy_feeder works:
  - Restrict it to only caching and serving round-0 entropy values.
    Previously it was dumping and rebuilding the cache when it
    encountered a different round number, but that's counterproductive
    because the next block is likely to be round 0 again (and thus would
    require flushing and rebuilding a second time).  Non-first-round
    blocks are now always cache misses, so that first round pulse blocks
    will be cache hits and not cause a double cache miss.
  - Increase the pulse feeder buffer size so that we can shift entropy
    values much less often.  We now use a sliding window inside data for
    the current active cache values, growing it towards the end of the
    array, and then memcpy entropy back to the beginning only once we
    hit the end (once every 30-something blocks rather than on every
    block).
- Add HF23 to testnet.